refactor: move languages types declarations into separate file
